refactor #23 color 파일 위치 변경
