--- lib/utils/colors.dart
@@ -1,28 +1,0 @@
-import "dart:ui";
-
-const white = Color(0xFFFFFFFF);
-const grey10 = Color(0xFFEBE7ED);
-const grey30 = Color(0xFFC5C1C7);
-const grey50 = Color(0xFF7A7681);
-const grey70 = Color(0xFF504B58);
-const grey80 = Color(0xFF38353E);
-const grey90 = Color(0xFF2E2B34);
-const grey100 = Color(0xFF262226);
-const grey110 = Color(0xFF161416);
-const grey120 = Color(0xFF1D1A1D);
-
-const trans = Color(0x80000000); //<- 투명도 50%
-const brandTrans = Color(0xCC996BFA); //<- 투명도 80%
-
-const brandColor = Color(0xFF996BFA);
-const brandDark = Color(0xFF7E66B0);
-const brandDisable = Color(0xFFA590D0);
-const tag = Color(0xFF3D2E5A);
-const redLight = Color(0xFFFFB9BB);
-const red = Color(0xFFFF5C62);
-const redDark = Color(0xFF4C3942);
-const redDarker = Color(0xFF2D1B1E);
-const blueLight = Color(0xFF9397FF);
-const blue = Color(0xFF5A61FF);
-const blueDark = Color(0xFF3C3A51);
-const blueDarker = Color(0xFF242345);
